Freeze version of urllib3 to prevent the pipeline failing
It started failing with error:
```
AttributeError: module 'urllib3.connectionpool' has no attribute 'VerifiedHTTPSConnection'
```

Most likely it was caused by installing newer available version, which no longer contains the VerifiedHTTPSConnection.
Also, there is another deprecation version coming from urllib3, reported from tests.

We have to bump the version of openapi generator to be able to bump the version of urllib3.
It will be a complex task, so for now we freeze the version of urllib3 as workaround solution.