BUGFIX: Get only docs/content/en from a PR
When we build the preview in pull_request_target action
we want to get only docs/content/en (the actual documentation files)
from the PR.
The rest of the repository can be used from master branch (and fork).
This is more secure as we don't allow users to inject malicious
commands in the building process (e.g. scripts in package.json files).