fix: failed fdw docker image build with postgres:12.15
There were several issues with the latest postgres:12.15. First, docker-compose build gooddata-fdw stopped working out of the box with the missing pg_config. I tried to fix it by adding libpq-dev, but it failed with an incompatible version of clang. The expected version is 15, but the installed is 16. Therefore, I decided to lock the base image to postgres:12.14-alpine which works.