PSDK-162: Fix tags of date dims
Put them all to a dedicated tag(folder).
Date dims are not put to the corresponding folder in our Analytical Designer.
Also, the same date dim can be created from multiple columns (shared date dims).
So the tagging is useless here now, we should think about it in the future.