fix(gooddata-sdk): add additional vis table type detection condition
In case a table with no row attributes is received, it is incorrectly
assumed to be a non-table ("non-pivot") visualization.

There doesn't seem to be a nice way to figure out whether a viz is table or not.
So for now we can try to parse it out from the visualizationUrl field if present.

JIRA: CQ-440