fix: allow specifying timeout for get_visualization
It was reported by customer that they would appreciate to have the timeout setting for get_visualization method. We agreed that it would be the best to have this setting for every exposed method. In this first iteration, it is only exposed for this specific method. Note, that we want to add it for other methods in the future, and also add global setting.

JIRA: EB-119