feat: add support for ExecutionContextAttributes
Reflects the changes done on the backend so that the FlightRPC function gets more information about the attributes used in the execution.

Also move to attrs for the class definitions in the area so that the serde is easier. In a future PR we might even move to cattrs for this, but not worth it now.

JIRA: CQ-577
risk: low