feat: improve resolving oneOf relationships
There was an issue resolving oneOf. A workaround was introduced some time ago, but it did not work for ExportDefinitionRequest. ExportDefinitionRequest was correctly matched, but it did not pass the assert, so it failed with the error that none oneOf could be used. Therefore, assert was extended for the use case when all model_kwargs are presented in oneof_class.openapi_types. This new condition and the old one should cover both cases.

JIRA: PSDK-209
risk: low