fix: release does not contain bumped version
Since rel/1.14.0 release branches do not contain the commit with version bump. The issue is in workflow that does not guarantee integrity. The workflow uses specific release branch now rel/x.y.z from the beginning and because it is protected branch it will not get deleted. Only patch release has different branch name – we do not want to persist it.

JIRA: PSDK-213
risk: low