feat: make the polling extension opt-in
By default, the FlexConnect will conform to the Arrow Flight RPC spec.
However, if an opt-in header is present, it will use the polling
extension used by GoodData.
This allows for things like query cancellation.
Ideally, we would use the PollFlightInfo from the Arrow Flight RPC
but unfortunately it is not yet available in PyArrow.

JIRA: CQ-1124
risk: low