feat: add on_execution_submitted to gooddata_pandas
Users can now provide a callback that will be called with the Execution
object as soon as it is available. This is useful for getting some of
the information available in the Execution object before the actual
data is loaded. The Execution object can also be used to cancel the
execution result operation.

JIRA: CQ-1387
risk: low